Security Parch

+++ Backend.API/DTOs/AdjustStockDto.cs
@@ -5,7 +5,7 @@ namespace Backend.API.DTOs;
 public class AdjustStockDto
 {
     [Required]
-    public int QuantityChange { get; set; }
+    public decimal QuantityChange { get; set; }
 
     [Required]
     public string Reason { get; set; } = string.Empty;
